# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Donation Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginBody {
  mobile: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  role: string;
  isActive: boolean;
  isSuspended?: boolean;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  aadharNumber?: string | null;
  /** @nullable */
  fatherHusbandName?: string | null;
  createdAt: string;
}

export interface LoginResponse {
  otpRequired: boolean;
  userId?: number;
  maskedEmail?: string;
  user?: User;
  token?: string;
}

export interface VerifyOtpBody {
  userId: number;
  /**
   * @minLength 6
   * @maxLength 6
   */
  otp: string;
}

export interface VerifyOtpResponse {
  user: User;
  token: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  super_admin: "super_admin",
  admin: "admin",
  collector: "collector",
  public: "public",
} as const;

export interface CreateUserBody {
  name: string;
  mobile: string;
  role: CreateUserBodyRole;
  password?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  aadharNumber?: string | null;
  /** @nullable */
  fatherHusbandName?: string | null;
}

export interface UpdateUserBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  isActive?: boolean | null;
  /** @nullable */
  isSuspended?: boolean | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  aadharNumber?: string | null;
  /** @nullable */
  fatherHusbandName?: string | null;
}

export interface Donation {
  id: number;
  donationId: string;
  name: string;
  mobile: string;
  amount: number;
  /** @nullable */
  purpose: string | null;
  /** @nullable */
  collectorId: number | null;
  /** @nullable */
  collectorName: string | null;
  hash: string;
  isVerified: boolean;
  createdAt: string;
}

export interface CreateDonationBody {
  name: string;
  mobile: string;
  amount: number;
  /** @nullable */
  purpose?: string | null;
}

export interface DonationListResponse {
  donations: Donation[];
  total: number;
  page: number;
  limit: number;
}

export interface VerificationResult {
  valid: boolean;
  donation: Donation;
  message: string;
}

export interface AnalyticsSummary {
  totalAmount: number;
  totalDonations: number;
  totalCollectors: number;
  totalDonors: number;
  todayAmount: number;
  todayDonations: number;
  avgDonation: number;
}

export interface DailyTotal {
  date: string;
  amount: number;
  count: number;
}

export interface CollectorStats {
  collectorId: number;
  collectorName: string;
  totalAmount: number;
  donationCount: number;
}

export interface AmountBucket {
  range: string;
  count: number;
  amount: number;
}

export interface AuditLog {
  id: number;
  /** @nullable */
  userId: number | null;
  /** @nullable */
  userName: string | null;
  action: string;
  /** @nullable */
  details: string | null;
  /** @nullable */
  ipAddress: string | null;
  createdAt: string;
}

export type ListDonationsParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  collector_id?: number | null;
  /**
   * @nullable
   */
  search?: string | null;
};

export type ListAuditLogsParams = {
  page?: number;
  limit?: number;
};