# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise Donation Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Verify mobile and password, then start OTP login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify a one-time password and issue a token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpResponse"
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List all donations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: collector_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of donations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationListResponse"

    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a new donation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDonationBody"
      responses:
        "201":
          description: Donation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get donation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteDonation
      tags: [donations]
      summary: Delete a donation (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /donations/{id}/verify:
    get:
      operationId: verifyDonation
      tags: [verify]
      summary: Verify donation QR code hash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        mobile:
          type: string
        password:
          type: string
      required:
        - mobile
        - password
    LoginResponse:
      type: object
      properties:
        otpRequired:
          type: boolean
        userId:
          type: integer
        maskedEmail:
          type: string
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - otpRequired
    VerifyOtpBody:
      type: object
      properties:
        userId:
          type: integer
        otp:
          type: string
          minLength: 6
          maxLength: 6
      required:
        - userId
        - otp
    VerifyOtpResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        isSuspended:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        aadharNumber:
          type: ["string", "null"]
        fatherHusbandName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - mobile
        - role
        - isActive
        - createdAt
...
    Donation:
      type: object
      properties:
        id:
          type: integer
        donationId:
          type: string
        name:
          type: string
        mobile:
          type: string
        amount:
          type: number
        purpose:
          type: ["string", "null"]
        collectorId:
          type: ["integer", "null"]
        collectorName:
          type: ["string", "null"]
        hash:
          type: string
        isVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - donationId
        - name
        - mobile
        - amount
        - purpose
        - collectorId
        - collectorName
        - hash
        - isVerified
        - createdAt
    CreateDonationBody:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        amount:
          type: number
        purpose:
          type: ["string", "null"]
      required:
        - name
        - mobile
        - amount
    DonationListResponse:
      type: object
      properties:
        donations:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - donations
        - total
        - page
        - limit
    VerificationResult:
      type: object
      properties:
        valid:
          type: boolean
        donation:
          $ref: "#/components/schemas/Donation"
        message:
          type: string
      required:
        - valid
        - donation
        - message
...