# lib/api-zod/src/generated/types/createDonationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Donation Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDonationBody {
  name: string;
  mobile: string;
  amount: number;
  /** @nullable */
  purpose?: string | null;
}