# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Donation Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface LoginResponse {
  otpRequired: boolean;
  userId?: number;
  maskedEmail?: string;
  user?: User;
  token?: string;
}