# lib/api-zod/src/generated/types/verifyOtpBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Donation Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyOtpBody {
  userId: number;
  /**
   * @minLength 6
   * @maxLength 6
   */
  otp: string;
}